Add private prop to Set
